backend/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
autobahn==23.1.2
Automat==24.8.0
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.3
daphne==4.1.2
decorator==5.2.1
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
executing==2.2.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
msgpack==1.1.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pyOpenSSL==25.0.0
python-decouple==3.8
redis==6.1.0
service-identity==24.2.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
wcwidth==0.2.13
whitenoise==6.9.0
zope.interface==7.2
